+++ collective/pdfpeek/dx/behavior.py
@@ -1,0 +1,10 @@
+from plone.directives import form
+from zope.interface import Interface
+
+
+class IPrimaryFieldPDFThumbnailsAndMetadata(form.Schema):
+    """Marker interface for extractable primary field"""
+
+
+class IPDFDataExtractionEnabled(Interface):
+    """ """